@@ -1,51 +1,85 @@
-<template>
-  <div id="app">
-    <app-header />
-    <keep-alive>
-      <router-view />
-    </keep-alive>
-  </div>
-</template>
-
-<script>
-import AppHeader from "@/components/AppHeader";
-import { request } from "@/api";
-
-export default {
-  components: { AppHeader },
-  mounted() {
-    this.getConfig();
-  },
-  methods: {
-    getConfig() {
-      request
-        .get("/configuration")
-        .then((response) => {
-          this.$store.dispatch("config/setConfig", response.data);
-        })
-        .catch((error) => {
-          throw new Error(error);
-        });
-    },
-  },
-};
+<script setup lang="ts">
+import { RouterLink, RouterView } from 'vue-router'
+import HelloWorld from './components/HelloWorld.vue'
 </script>
 
-<style lang="scss">
-*,
-*::before,
-*::after {
-  box-sizing: border-box;
+<template>
+  <header>
+    <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
+
+    <div class="wrapper">
+      <HelloWorld msg="You did it!" />
+
+      <nav>
+        <RouterLink to="/">Home</RouterLink>
+        <RouterLink to="/about">About</RouterLink>
+      </nav>
+    </div>
+  </header>
+
+  <RouterView />
+</template>
+
+<style scoped>
+header {
+  line-height: 1.5;
+  max-height: 100vh;
 }
-body {
-  margin: 0;
-  background-color: #040302;
-  color: $text-color;
+
+.logo {
+  display: block;
+  margin: 0 auto 2rem;
 }
-#app {
-  font-family: "Avenir", Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  color: #f5f5f4;
+
+nav {
+  width: 100%;
+  font-size: 12px;
+  text-align: center;
+  margin-top: 2rem;
+}
+
+nav a.router-link-exact-active {
+  color: var(--color-text);
+}
+
+nav a.router-link-exact-active:hover {
+  background-color: transparent;
+}
+
+nav a {
+  display: inline-block;
+  padding: 0 1rem;
+  border-left: 1px solid var(--color-border);
+}
+
+nav a:first-of-type {
+  border: 0;
+}
+
+@media (min-width: 1024px) {
+  header {
+    display: flex;
+    place-items: center;
+    padding-right: calc(var(--section-gap) / 2);
+  }
+
+  .logo {
+    margin: 0 2rem 0 0;
+  }
+
+  header .wrapper {
+    display: flex;
+    place-items: flex-start;
+    flex-wrap: wrap;
+  }
+
+  nav {
+    text-align: left;
+    margin-left: -1rem;
+    font-size: 1rem;
+
+    padding: 1rem 0;
+    margin-top: 1rem;
+  }
 }
 </style>
